Modificación de UX para las funcionalidades de admin
En el panel de admin se mejoraron las UX para las funcionalidades de cambiar contraseña y eliminar usuario

+++ app/Http/Controllers/AdminRESTController.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\Controller;
 use Auth;
 use DB;
 use Hash;
+use Validator;
 class AdminRESTController extends Controller
 {
 
@@ -44,35 +45,53 @@ class AdminRESTController extends Controller
 
   public function changePassword(Request $request){
 
-    $validatedData = $request->validate([
-        'id' => 'bail|required|exists:users,id',
-        'password' => 'min:6|required_with:password_confirmation|same:password_confirmation',
+    $input = $request->all();
+    $rules = array(
+        'userId'        => 'required|exists:users,id',
+        'new_password'  => 'min:6|required_with:password_confirmation|same:password_confirmation',
         'password_confirmation' => 'min:6'
-    ]);
+        );
+    $messages = array(
+        'required'          => 'El campo :attribute es requerido.',
+        'exists'            => 'El usuario ingresado no existe.',
+        'required_with'     => 'Las contraseñas deben coincidir.',
+        'same'              => 'Las contraseñas deben coincidir.',
+        'min'               => 'La contraseña debe poseer un mínimo de :min caracteres.'
+    );
 
-    $id = $request['id'];
-    $user = User::where('id', $id)->first(); 
-    $user->password = Hash::make($request['password']);
-    $user->save();
+    $validator = Validator::make($input,$rules,$messages);
+    if ($validator->passes()){
+        $id = $input['userId'];
+        $user = User::where('id', $id)->first();
+        $user->password = Hash::make($input['new_password']);
+        $user->save();
+    }
 
-    return redirect('panel/admin-list');
-  }
+    return $validator->messages();
+}
 
   public function deleteUser(Request $request){
 
-    $validatedData = $request->validate([
-        'id' => 'required|exists:users,id'
-    ]);
+    $input = $request->all();
+    $rules = array(
+        'userId'        => 'required|exists:users,id'
+        );
+    $messages = array(
+        'required'          => 'El campo :attribute es requerido.',
+        'exists'            => 'El usuario ingresado no existe.'
+    );
 
-    $id = $request['id'];
-    if($id == Auth::id()){//Cannot delete myself
-
-        return redirect('panel/admin-list');
+    $validator = Validator::make($input,$rules,$messages);
+    if ($validator->passes()){
+        $id = $request['userId'];
+        if($id == Auth::id()){//Cannot delete myself
+            return -1;
+        }
+        $user = User::where('id', $id)->first();
+        $user->delete();
     }
-    $user = User::where('id', $id)->first();
-    $user->delete();
 
-    return redirect('panel/admin-list');
+    return $validator->messages();
   }
 
     /**
